Add User::Id extractor for use in all routes
This was a huge headache to add, mainly because of the incompatibility
between `dyn Future` and `async` functions.  `dyn Future` by default
requires a `'static` lifetime, while `async` functions use an anonymous
lifetime on the return type of the Future.  This was done in order to
call the `status` function on the blacklist, but in the long-term it
will likely make more sense to do that check in middleware, where we can
use async functions more easily.